feat: add iOS/Android/Web platform support and fix Flutter build errors
- Generate iOS, Android, and Web platform directories via flutter create
- Configure iOS Info.plist with location, camera, photo permissions
- Add Google Maps SDK setup in AppDelegate.swift
- Add Podfile with iOS 14.0 minimum deployment target
- Replace non-existent signature_pad_widget with hand_signature package
- Fix missing AuthEvent import in app.dart
- Fix distanceFilter type in location_service.dart
- Fix TabBar onTap callback type in invoice_list_page.dart
- Fix widget_test.dart to match actual app structure

+++ mobile/lib/core/services/location_service.dart
@@ -44,7 +44,7 @@ class LocationService {
 
   Stream<Position> getPositionStream({
     int intervalMs = 5000,
-    double distanceFilter = 10.0,
+    int distanceFilter = 10,
   }) {
     return Geolocator.getPositionStream(
       locationSettings: LocationSettings(
